grafana-ui: lazyify DnD, FileDropzone, and keybinding
