+++ packages/grafana-ui/src/unstable.ts
@@ -34,6 +34,8 @@ export type {
   SignatureInformation,
   SignatureParameter,
 } from './components/CodeMirror/types';
+export { useDragAndDrop } from './components/DragAndDrop/useDragAndDrop';
+export type { DragAndDropComponents, DragAndDropModule } from './components/DragAndDrop/useDragAndDrop';
 export { getQueryFieldConfig } from './components/QueryFieldConfig/queryFieldConfig';
 export type { QueryFieldConfig, QueryFieldConfigOptions } from './components/QueryFieldConfig/queryFieldConfig';
 export { TableNG } from './components/Table/TableNG/TableNG';
